Rename group numbers
